Add -set_exit_status flag
Add support for the -set_exit_status flag to goconst, inspired by the
same flag on the prealloc tool [1]. I updated the help message
accordingly. The exit status is set to `2` to distinguish it from an
non-recoverable error, indicated by an exit status of `1`.

I implemented this by hvaing the `run` function return an additional
value indicating if anyIssues were found. `run` gets this information
from `printOutput` and returns it directly. This is needed because
`printOutput` does some additional processing of the duplicates found
(namely checking if there are enough occurances of a given constant).

If there is any error in the program, this always takes precedence over
issues found by goconst.

--
1. https://github.com/alexkohler/prealloc